Improve reported metrics in BulkLoadDemo (#27)
BulkLoadDemo currently shows VRAM allocations, but those sizes may be larger than the amount of data actually loaded, e.g. texture VRAM also includes padding. And so using those numbers to calculate the bandwidth is incorrect.

This change fixes it so that the demo instead reports the data loaded and decompressed by DirectStorage, which will make the bandwidth more accurate.